Add recursive directory support to the "depend_on" option
When the depend_on option is given a path that references a directory,
the Dependable module will include the contents of that directory, and
recursively walk the directory for other directories (& their contents)
until there are no directories left to traverse into.

This means that, given the following content directory:

* content/lib/foo/1/2/foo.js
* content/lib/bar/1/2/bar.js
* content/lib/bar/1/2/baz.js
* content/lib/foobarbaz.js

And given that `depend_on` is provided with "lib/", all of the above
paths will be tracked as dependencies of a webpack entry point.